+++ index.ts
@@ -1,1 +1,4 @@
 export * from './src/BrazeMessageComponent';
+export * from './src/logic/BrazeMessages';
+export * from './src/logic/NullBrazeMessages';
+export * from './src/logic/LocalMessageCache';
