+++ src/SpecialEditionBanner/canRender.tsx
@@ -1,0 +1,1 @@
+export const COMPONENT_NAME = 'SpecialEditionBanner';
