+++ src/TheGuardianIn2020Banner/canRender.ts
@@ -1,0 +1,8 @@
+import type { BrazeMessageProps } from './index';
+
+export const COMPONENT_NAME = 'TheGuardianIn2020Banner';
+
+export const canRender = (brazeMessageProps: BrazeMessageProps): boolean => {
+    const { header, body } = brazeMessageProps;
+    return Boolean(header && body);
+};
