+++ src/logic/BrazeMessages.tsx
@@ -3,7 +3,16 @@
 import type appboy from '@braze/web-sdk-core';
 import { MessageCache } from './LocalMessageCache';
 import type { SlotName } from './types';
-import { canRenderBrazeMsg } from '../BrazeMessageComponent';
+import {
+    COMPONENT_NAME as APP_BANNER_NAME,
+    canRender as appBannerCanRender,
+} from '../AppBanner/canRender';
+import {
+    COMPONENT_NAME as G_IN_2020_BANNER_NAME,
+    canRender as gIn2020BannerCanRender,
+} from '../TheGuardianIn2020Banner/canRender';
+import { COMPONENT_NAME as DIGITAL_SUBSCRIBER_APP_BANNER } from '../DigitalSubscriberAppBanner/canRender';
+import { COMPONENT_NAME as SPECIAL_EDITION_BANNER } from '../SpecialEditionBanner/canRender';
 
 export type Extras = Record<string, string>;
 export type ErrorHandler = (error: Error, identifier: string) => void;
@@ -12,6 +21,26 @@ interface BrazeMessagesInterface {
     getMessageForBanner: () => Promise<BrazeMessage>;
     getMessageForEndOfArticle: () => Promise<BrazeMessage>;
 }
+
+const COMPONENT_CAN_RENDER_MAPPINGS: Record<
+    string,
+    (brazeMessageProps: Record<string, unknown>) => boolean
+> = {
+    [APP_BANNER_NAME]: appBannerCanRender,
+    [DIGITAL_SUBSCRIBER_APP_BANNER]: appBannerCanRender,
+    [SPECIAL_EDITION_BANNER]: appBannerCanRender,
+    [G_IN_2020_BANNER_NAME]: gIn2020BannerCanRender,
+};
+
+export const canRenderBrazeMsg = (msgExtras: Extras | undefined): boolean => {
+    if (!msgExtras) {
+        return false;
+    }
+    if (!COMPONENT_CAN_RENDER_MAPPINGS[msgExtras.componentName]) {
+        return false;
+    }
+    return COMPONENT_CAN_RENDER_MAPPINGS[msgExtras.componentName](msgExtras);
+};
 
 const generateId = (): string => `${Math.random().toString(16).slice(2)}-${new Date().getTime()}`;
